Default to 32

+++ hdi/dimensionality_reduction/gradient_descent_tsne_texture.h
@@ -55,7 +55,7 @@ namespace hdi {
     \author Nicola Pezzotti
     */
 
-    template <typename scalar = float, typename sparse_scalar_matrix = std::vector<hdi::data::MapMemEff<std::uint32_t, float> >, typename unsigned_integer = std::uint64_t, typename integer = std::int64_t>
+    template <typename scalar = float, typename sparse_scalar_matrix = std::vector<hdi::data::MapMemEff<std::uint32_t, float> >, typename unsigned_integer = std::uint32_t, typename integer = std::int32_t>
     class GradientDescentTSNETexture {
     public:
 #ifndef __APPLE__
