+++ hdi/dimensionality_reduction/sparse_tsne_user_def_probabilities.h
@@ -52,7 +52,7 @@ namespace hdi{
       Implementation of the tSNE algorithm with sparse and user-defined probabilities
       \author Nicola Pezzotti
     */
-    template <typename scalar = float, typename sparse_scalar_matrix = std::vector<hdi::data::MapMemEff<std::uint32_t, float> >, typename unsigned_integer = std::uint64_t, typename integer = std::int64_t>
+    template <typename scalar = float, typename sparse_scalar_matrix = std::vector<hdi::data::MapMemEff<std::uint32_t, float> >, typename unsigned_integer = std::uint32_t, typename integer = std::int32_t>
     class SparseTSNEUserDefProbabilities{
     public:
       typedef scalar scalar_type;
